+++ mdfrier/src/ratatui.rs
@@ -329,7 +329,7 @@ pub fn render_line<T: Theme>(md_line: MdLine, theme: &T) -> (Line<'static>, Vec<
             if let Some(link) = node.get_source_content() {
                 tags.push(Tag::Link(line_spans.len(), link));
             } else {
-                log::warn!("node has LinkURL but no source_content");
+                log::warn!("node has LinkURL but no source_content: {}", node.content);
                 tags.push(Tag::Link(
                     line_spans.len(),
                     SourceContent::from(node.content.as_ref()),